Rework the docs.rb script as a Rake task
